requirements.txt
fastapi
uvicorn
sqlmodel
alembic
psycopg2-binary
pydantic
pydantic-settings
playwright
requests
feedparser
langchain
langchain-openai
chromadb
streamlit
apscheduler
pyyaml
beautifulsoup4
apify-client
langchain-chroma
fastapi
uvicorn
